@@ -6,7 +6,7 @@ readme = "../README.md"
 repository = "https://github.com/OpenByteDev/burnt-sushi"
 license = "MIT"
 authors = ["OpenByte <development.openbyte@gmail.com>"]
-edition = "2021"
+edition = "2024"
 keywords = ["spotify", "adblocker", "windows", "blocker"]
 
 [dependencies]
